make SparseArrays a weak dependency
